+++ src/main/java/pascal/taie/analysis/pta/plugin/taint/ParamSourcePoint.java
@@ -30,7 +30,7 @@ import java.util.Comparator;
 /**
  * A {@code ParamSourcePoint} is a parameter of a method.
  */
-public record ParamSourcePoint(JMethod sourceMethod, IndexRef indexRef)
+record ParamSourcePoint(JMethod sourceMethod, IndexRef indexRef, ParamSource source)
         implements SourcePoint {
 
     private static final Comparator<ParamSourcePoint> COMPARATOR =